+++ CHANGELOG.md
@@ -5,7 +5,30 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.28.0] - 2022-01-
+## [0.29.0] - Unreleased
+
+### Removed
+
+- `showTagsControl`. This was used to hide the button that opened the Tags List
+  Overlay
+
+- The Tag List Overlay is now part of a new Asset List Overlay. See below.
+
+### Added
+
+- `api.fetchPlacemarksByLocation()`
+
+- Asset List Overlay. This replaces the Tag List Overlay and provides the
+  ability to also search for Placemarks.
+
+- `showSearchControl`. This can be used to hide the search icon that opens the
+  Asset List overlay
+
+## [0.28.2] - 2021-12-06
+
+### Updated
+
+- Dependency updates
 
 ## [0.27.0] - 2021-11-08
 
@@ -350,7 +373,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `onTagUpdate` was not called from `createAPI`
 
-- If `showTagsControl` was false, all control tags would be
+- Bug: If `showTagsControl` was false, all control tags would be
   hidden on the map
 
 - The map background image would fail to load unless the user